Changed `util.misc.run_in_terminal` to first look for tmux, and then X11
This is needed because looking the environment variable `XAUTHORITY` is not enough to detect X11.  Specifically it is not set on my system.  The existence of the file `~/.Xauthority` is taken as an indication of running under X11.  But this file will still be present when running tmux, hence the changed order of detection.